+++ src/main.jsx
@@ -1,8 +1,8 @@
 // src/main.jsx
 import React from "react";
 import ReactDOM from "react-dom/client";
+import AuthProvider from "./context/AuthProvider";
 import App from "./App";
-import { AuthProvider } from "./context/AuthContext";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./index.css";
