+++ src/pages/Dashboard.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../hooks/useAuth";
 
 export default function Dashboard() {
     const { getCurrentUser, logout } = useAuth();
